Fix sort command bug and User Guide documentation.
Currently, The sort command has a bug where the ordering of persons with null
appointments will change each time the function is called. This bug fix fixes
the functionality and causes people will null appointments to be ordered in an
arbitrary order based on the previous way the list was sorted.

Edits were also made to the user guide to provide further clarification to the
ordering of peoples names when sort appointment is called.